requirements.txt
Django==1.10.4
django-filter==1.0.1
djangorestframework==3.5.3
elasticsearch==5.0.1
Markdown==2.6.7
PyJWT==1.4.2
urllib3==1.19.1
psycopg2==2.6.2
django-cors-headers==2.0.0
certifi==2016.9.26
architect==0.5.5

